Implement LinkTent, fix LinkTree, fix NoTentForTree
Implemented the LinkTent Case rule, and applied several elements of the logic I used there to improve reliability to LinkTree. Additionally, in the course of testing LinkTent and LinkTree, I noticed a bug with NoTentForTree not checking for empty tiles correctly, so I included a fix for that as well.