Halfway - Menu Improvements + Providing Help to User
It is an unfinished version. With talking to the group leader, we decide to use a web page to replace the Pop-up window dialog. So far, the web page hasn't set up yet, I just put Github legup link here, waiting to have a real dialog web page in the future.