Color blind support (#470) (#471)
* Color blind support (#470)

* new CB resources

a new folder of assets that are red and blue not red and green

* Added the color blind preference setting

Added the color blind preference setting in the preference page and the user preference map.

* atomic

i changed atomic Rule because i forgot to change it last week

* Modifies loadimage() function

The loadimage() function now checks if the user selects color blind mode.

* changes grid colors in compliance with colorblind preference

i changed the element view to update colors. now all thats left is changing the proof tree.

* Implements color blind settings on the proof tree

---------

Co-authored-by: Jason Pu <jason.pu19@outlook.com>

* Color blind support (#483)

* new CB resources

a new folder of assets that are red and blue not red and green

* Added the color blind preference setting

Added the color blind preference setting in the preference page and the user preference map.

* atomic

i changed atomic Rule because i forgot to change it last week

* Modifies loadimage() function

The loadimage() function now checks if the user selects color blind mode.

* changes grid colors in compliance with colorblind preference

i changed the element view to update colors. now all thats left is changing the proof tree.

* Implements color blind settings on the proof tree

* Bug fix

* colorBlind -> colorblind

* Error fixed but good

---------

Co-authored-by: Jason Pu <jason.pu19@outlook.com>
Co-authored-by: Charles Tian <46334090+charlestian23@users.noreply.github.com>

---------

Co-authored-by: Jason Pu <jason.pu19@outlook.com>
Co-authored-by: Charles Tian <46334090+charlestian23@users.noreply.github.com>