Generalized max cases
Max cases is generalized for each case rule. Can be altered from the original 10 to whichever you choose. I input 5 for finish room and 2 for black or white case rule in Nurikabe but can be done for case rules in other puzzles. AutoCaseRuleCommand was changed as well to display message when too many cases of the case rule is generated.