Tree Tent Test Suite (#775)
* TentForTree test cases

One valid case of connecting a tree to a tent
One invalid case of connecting a tree to a tent when there is another valid tent

* Incomplete test for already connected tree

* Test for Fill in Row Case Rule

Looks like the case rule was implemented incorrectly :/

* Fill in Row Test Cases

Actually case generator works properly

* Fill in Row Test Cases

When the clue is that there are 0 tents, the case returned should be equivalent to FinishWithGrass.

* Comments

* Fix TreeTent Puzzle Editor

* LinkTentCaseRule Four Trees Test Case

Expected to create four cases, each one connecting a line between the tree in the center and one of the four surrounding trees.

* LinkTentCaseRule Various Test Cases

With one tree, one case is created.
With no trees, no cases are created.
With diagonal trees, no cases are created.

* TreeTent Puzzle Editor Fix

Removed double-calling setCell due to custom and super mouseReleased

* LinkTreeCaseRule Test Cases

Similar test cases to LinkTentCaseRule
- Make sure diagonal tents have no bearing on cases
- Make sure no tents around the tree fails
- One tent should connect to the correct tree
- Two tents should create two cases
- No other cases to check, as there cannot be more than two tents around a tree; otherwise, the tents would touch

* Reset puzzle

* Improved FillInRowCaseRule

Rewrote recursive generate board function to be slightly more optimized

* Fix FillInRowCaseRule

* Implement Copy Clue function

Fixes a bug where you couldn't apply the FillInRow Case Rule on a child node

* Allow Importing Puzzle-Specific Puzzle Elements

When checking the proof tree, the Puzzle Importer only checks for changed cells. However, for TreeTent, the user can create lines, which cannot be saved as a "cell" object. Now, an importer class can override the getImporterElements function to specify what types of changes the board may experience.

* Fix Puzzles Uneven Sides

Simple mistake in calculation of index

* Check Rule on Columns

Adjust all simple cases to also check the central column rule

* Change test puzzle names

* FillInRowCaseRule Test

Checks that the rule works on an empty 5x5 board

* TreeForTent tests and TentForTree tests

* Minor formatting changes

* Fix boards with uneven side lengths

* Comments

* Standardizing Comments

* Standardizing Comments

All Direct Rules

* Fix Build Issues

The Ubuntu and Checkstyle Autocompilers have trouble opening this file for some reason

* Comment out problematic file

* Uncommenting problematic test

* Update file name

* Commented out Star Battle test

---------

Co-authored-by: Charles Tian <46334090+charlestian23@users.noreply.github.com>
Co-authored-by: charlestian23 <charles.tian23@gmail.com>